# Use Ninja for BoringSSL if available