--- a/src/soter/boringssl/soter.mk
+++ b/src/soter/boringssl/soter.mk
@@ -36,6 +36,10 @@ CRYPTO_ENGINE_LDFLAGS += -lcrypto -ldecrepit -lpthread
 SOTER_ENGINE_CMAKE_FLAGS += -DCMAKE_BUILD_TYPE=Release
 SOTER_ENGINE_CMAKE_FLAGS += -DCMAKE_C_FLAGS="-fpic"
 
+ifneq ($(NINJA),)
+SOTER_ENGINE_CMAKE_FLAGS += -G Ninja
+endif
+
 ifdef IS_LINUX
 RENAME_BORINGSSL_SYMBOLS = yes
 endif
@@ -60,7 +64,11 @@ $(BIN_PATH)/boringssl/crypto/libcrypto.a $(BIN_PATH)/boringssl/decrepit/libdecre
 	@mkdir -p $(BIN_PATH)/boringssl/stage-1
 	@cd $(BIN_PATH)/boringssl/stage-1 && \
 	 $(CMAKE) $(SOTER_ENGINE_CMAKE_FLAGS) $(abspath third_party/boringssl/src)
+ifeq ($(NINJA),)
 	@$(MAKE) -C $(BIN_PATH)/boringssl/stage-1 crypto decrepit
+else
+	@$(NINJA) -C $(BIN_PATH)/boringssl/stage-1 crypto decrepit
+endif
 	@mkdir -p $(BIN_PATH)/boringssl/crypto $(BIN_PATH)/boringssl/decrepit
 	@cp $(BIN_PATH)/boringssl/stage-1/crypto/libcrypto.a     $(BIN_PATH)/boringssl/crypto/libcrypto.a
 	@cp $(BIN_PATH)/boringssl/stage-1/decrepit/libdecrepit.a $(BIN_PATH)/boringssl/decrepit/libdecrepit.a
@@ -78,7 +86,11 @@ ifeq ($(RENAME_BORINGSSL_SYMBOLS),yes)
 	     -DBORINGSSL_PREFIX=$(SOTER_BORINGSSL_PREFIX) \
 	     -DBORINGSSL_PREFIX_SYMBOLS=../symbols.txt \
 	     $(abspath third_party/boringssl/src)
+ifeq ($(NINJA),)
 	@$(MAKE) -C $(BIN_PATH)/boringssl/stage-2 crypto decrepit
+else
+	@$(NINJA) -C $(BIN_PATH)/boringssl/stage-2 crypto decrepit
+endif
 	@mkdir -p $(BIN_PATH)/boringssl/crypto $(BIN_PATH)/boringssl/decrepit
 	@cp $(BIN_PATH)/boringssl/stage-2/crypto/libcrypto.a     $(BIN_PATH)/boringssl/crypto/libcrypto.a
 	@cp $(BIN_PATH)/boringssl/stage-2/decrepit/libdecrepit.a $(BIN_PATH)/boringssl/decrepit/libdecrepit.a