Ninja build tool significantly reduces compilation time for BoringSSL in default configuration.

This is mostly due to the fact that it performs parallel builds by default but there are other gains (Makefiles that CMake generates are not that optimal). `make -j4` is also a viable option, but it often ends up brittle.

<details>
<summary>Timings for my machine</summary>

As you can see, there is some gain in wall clock time (real), resulting in shorter builds. However, this comes mostly from better parallelization if you look at the user time.

### `time make ENGINE=boringssl`

master:

```
real	4m30.623s
user	3m32.624s
sys	0m33.530s
```

with ninja:

```
real	2m47.085s
user	4m47.580s
sys	0m34.020s
```

### `time emmake make wasmthemis`

master:

```
real	2m59.856s
user	2m14.338s
sys	0m29.650s
```

with ninja:

```
real	1m57.756s
user	3m13.189s
sys	0m33.403s
```

</details>

<details>
<summary>Timings for GitHub Actions</summary>

It's always nice to see builds run a bit quicker.

master:

| Test set    | Job                 | Step                          | Time   |
| --          | --                  | --                            | --     |
| Themis Core | Unit tests (ubuntu) | Build Themis Core (BoringSSL) | 3m 09s |
| Themis Core | Unit tests (macos)  | Build Themis Core (BoringSSL) | 6m 57s |
| Themis Core | Memory leaks        | Build Themis Core             | 4m 22s |
| WasmThemis  | Unit tests (16.x)   | Build WasmThemis              | 2m 34s |

with ninja:

| Test set    | Job                 | Step                          | Time   |
| --          | --                  | --                            | --     |
| Themis Core | Unit tests (ubuntu) | Build Themis Core (BoringSSL) | 2m 33s |
| Themis Core | Unit tests (macos)  | Build Themis Core (BoringSSL) | 2m 26s |
| Themis Core | Memory leaks        | Build Themis Core             | 2m 48s |
| WasmThemis  | Unit tests (16.x)   | Build WasmThemis              | 1m 45s |

</details>

## Checklist

- [x] Change is covered by automated tests
- [x] Benchmark results are attached (if applicable)
- [x] The [coding guidelines] are followed
- [x] Public API has proper documentation
- [x] Changelog is updated

[coding guidelines]: https://github.com/cossacklabs/themis/blob/master/CONTRIBUTING.md